feat: restructure package
패키지 구조 변경

Resolves:
Related to:
