+++ src/main/java/com/service/dida/DidaApplication.java
@@ -2,8 +2,10 @@ package com.service.dida;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
+@EnableJpaAuditing
 public class DidaApplication {
 
     public static void main(String[] args) {
