+++ src/main/java/com/service/dida/domain/comment/Comment.java
@@ -1,8 +1,8 @@
-package com.service.dida.comment;
+package com.service.dida.domain.comment;
 
-import com.service.dida.common.BaseEntity;
-import com.service.dida.post.Post;
-import com.service.dida.user.User;
+import com.service.dida.global.common.BaseEntity;
+import com.service.dida.domain.post.Post;
+import com.service.dida.domain.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
