+++ src/main/java/com/service/dida/domain/user/Role.java
@@ -1,9 +1,9 @@
-package com.service.dida.user;
+package com.service.dida.domain.user;
 
 public enum Role {
 
-    GENERAL("일반 회원"),
-    MANAGER("관리자");
+    ROLE_USER("일반 회원"),
+    ROLE_MANAGER("관리자");
 
     private final String name;
