+++ src/main/java/com/service/dida/global/config/exception/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.service.dida.config.exception;
+package com.service.dida.global.config.exception;
 
 public interface ErrorCode {
     String getErrorCode();
