+++ src/main/java/com/service/dida/global/config/response/ExceptionResponse.java
@@ -1,8 +1,10 @@
-package com.service.dida.config.response;
+package com.service.dida.global.config.response;
 
+import com.service.dida.global.config.exception.BaseException;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
+import org.springframework.http.ResponseEntity;
 
 @Getter
 public class ExceptionResponse {
